fix(change_stream): do not check isGetMore if error[mongoErrorContextSymbol] is undefined
Fixes NODE-1494
